refactor と 整理
